parser/requirements.txt
python-dotenv
requests
pandas
PyGithub
sqlalchemy
psycopg2